Updated README and Parser
Fixed the parser as it wasn't properly handling the creation of binary
and unary operators.

Added a multiple file processing phase for the main executable, this is
still broken as we're emitting all the individual files instead of just
the final AST. #TODO

Updated README with instructions on compiling the produced code to a
native executable.